feature/#524 [iOS/Android] language setting implementation
## Overview (Required)
- Implementation of language setting

## Links
- https://github.com/DroidKaigi/conference-app-2021/issues/524

## Other
I haven't checked how it works after the conflict is resolved yet, so I will check later.
It seems that the action after `catchToEffect()` is not called as well as the `addFavorite` part.
I only added the main thread specification once because it was crashing if I didn't process it in the main thread, but if anyone knows the cause of the problem, it would be great if you could advise me.